Reduced advance range, modified breakable crate nav, added test squad, and added route to kitchen from main hall.
Added a bunch of info_notpad s with the name "notes"

This is confusing me, sometimes they fight good, other times not. Most of the time they move up well, but that fact there's one door for most of the hallways is causing problems when they flank. Lowering advance range is helping a little.

Three soldiers is alot in the wrong situation. Might be OK for later squads. Chargers behave well and fun. Officer plus Charger is deadly. Starting with Two Grunts is nice and simple. Depends how intense lobby combat gets.